Implementando o resto do código da subtração e soma. O teste que funfa!

+++ src/main/Calculator.java
@@ -3,11 +3,11 @@ package main;
 public class Calculator {
 	private double resultado;
 
-	public int soma(int a, int b) {
+	public double soma(double a, double b) {
 		return a + b;
 	}
 
-	public int sub(int a, int b) {
+	public double sub(double a, double b) {
 		return a - b;
 	}
 }
